Successfully refactored Express middleware for S3 Routers
